Add files via upload
The Hop-by-Hop Options Header is taken into consideration by several network operators, as a valuable container for carrying the information facilitating the introduction of new services. However, currently the HBH Options Header is not yet actually utilised in the operators' networks. This two documents outline the reasons why it is rarely utilized in current networks, and describe the processing of the Hop-by-Hop Options Header in today's routers in the aspects of standards specification, common implementations, and default operations.